flip bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -5,7 +5,7 @@
 * @n: int n
 * Return: counter
 */
-int countbits(unsigned int n)
+int countbits(unsigned long int n)
 {
 	unsigned int count = 0;
